Add: test head and raytracing test
