+++ disc0ver-Engine/disc0ver-Engine/disc0ver-Engine.vcxproj
@@ -161,6 +161,8 @@
     <ClCompile Include="render\shader.cpp" />
     <ClCompile Include="render\texture.cpp" />
     <ClCompile Include="render\transform.cpp" />
+    <ClCompile Include="test\test_master.cpp" />
+    <ClCompile Include="test\test_rayTracing.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="engine\light\light.h" />
@@ -182,6 +184,7 @@
     <ClInclude Include="render\shader.h" />
     <ClInclude Include="render\texture.h" />
     <ClInclude Include="render\transform.h" />
+    <ClInclude Include="test\test_total.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
