+++ disc0ver-Engine/disc0ver-Engine/imgui.ini
@@ -4,7 +4,12 @@ Size=400,400
 Collapsed=0
 
 [Window][Another Window]
-Pos=45,72
-Size=198,155
+Pos=48,34
+Size=440,160
+Collapsed=0
+
+[Window][Raytracing Properties]
+Pos=11,18
+Size=402,125
 Collapsed=0
